+++ packages/capacitor/src/index.ts
@@ -1,2 +1,2 @@
+export { AddPluginSchematicSchema } from './schematics/add-plugin/schema';
 export { CapacitorSchematicSchema } from './schematics/capacitor-project/schema';
-export { runCapacitorCommand } from './utils/capacitor-command';