+++ packages/capacitor/CHANGELOG.md
@@ -11,6 +11,7 @@
 - add Capacitor configs to frontend application
 - add or update `package.json` in project folder when generating a Capacitor project
 - add builder configurations for Nx Console
+- add `add-plugin` schematic for adding Capacitor plugins
 
 # BREAKING CHANGES